chore(ecs): Rename check, change status extended and change test global variables to uppercase
